refactor(todos|todo): move to components
